fix(core): cleanup menu handling irregularities
[no changelog]
